backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
neo4j==5.14.1
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
httpx==0.25.2
pandas==2.1.3
python-multipart==0.0.6
aiofiles==23.2.1